AppNav Funcionando | TelaAddCtt funcionando(ao clicar no addButton na TelaDiscagem) | Teclado Numerico Funcionando
