+++ .idea/deploymentTargetSelector.xml
@@ -13,9 +13,6 @@
         </DropdownSelection>
         <DialogSelection />
       </SelectionState>
-      <SelectionState runConfigName="Discagem">
-        <option name="selectionMode" value="DROPDOWN" />
-      </SelectionState>
       <SelectionState runConfigName="TelaDiscagem">
         <option name="selectionMode" value="DROPDOWN" />
       </SelectionState>
